+++ carp_mobile_sensing/lib/probes/sensors/sensor_datum.dart
@@ -9,8 +9,8 @@ part of sensors;
 /// A [Datum] that holds acceleration data collected from the native accelerometer on the phone.
 @JsonSerializable(fieldRename: FieldRename.snake, includeIfNull: false)
 class AccelerometerDatum extends CARPDatum {
-  static CARPDataFormat CARP_DATA_FORMAT = new CARPDataFormat(
-      NameSpace.CARP_NAMESPACE, ProbeRegistry.ACCELEROMETER_MEASURE);
+  static CARPDataFormat CARP_DATA_FORMAT =
+      new CARPDataFormat(NameSpace.CARP_NAMESPACE, ProbeRegistry.ACCELEROMETER_MEASURE);
 
   /// Acceleration force along the x axis (including gravity) measured in m/s^2.
   double x;
@@ -21,11 +21,9 @@ class AccelerometerDatum extends CARPDatum {
   /// Acceleration force along the z axis (including gravity) measured in m/s^2.
   double z;
 
-  AccelerometerDatum({this.x, this.y, this.z})
-      : super(includeDeviceInfo: false);
+  AccelerometerDatum({this.x, this.y, this.z}) : super(includeDeviceInfo: false);
 
-  factory AccelerometerDatum.fromJson(Map<String, dynamic> json) =>
-      _$AccelerometerDatumFromJson(json);
+  factory AccelerometerDatum.fromJson(Map<String, dynamic> json) => _$AccelerometerDatumFromJson(json);
   Map<String, dynamic> toJson() => _$AccelerometerDatumToJson(this);
 
   CARPDataFormat getCARPDataFormat() => CARP_DATA_FORMAT;
@@ -36,8 +34,8 @@ class AccelerometerDatum extends CARPDatum {
 /// A [Datum] that holds rotation data collected from the native gyroscope on the phone.
 @JsonSerializable(fieldRename: FieldRename.snake, includeIfNull: false)
 class GyroscopeDatum extends CARPDatum {
-  static CARPDataFormat CARP_DATA_FORMAT = new CARPDataFormat(
-      NameSpace.CARP_NAMESPACE, ProbeRegistry.GYROSCOPE_MEASURE);
+  static CARPDataFormat CARP_DATA_FORMAT =
+      new CARPDataFormat(NameSpace.CARP_NAMESPACE, ProbeRegistry.GYROSCOPE_MEASURE);
 
   /// Rate of rotation around the x axis measured in rad/s.
   double x;
@@ -50,8 +48,7 @@ class GyroscopeDatum extends CARPDatum {
 
   GyroscopeDatum({this.x, this.y, this.z}) : super(includeDeviceInfo: false);
 
-  factory GyroscopeDatum.fromJson(Map<String, dynamic> json) =>
-      _$GyroscopeDatumFromJson(json);
+  factory GyroscopeDatum.fromJson(Map<String, dynamic> json) => _$GyroscopeDatumFromJson(json);
   Map<String, dynamic> toJson() => _$GyroscopeDatumToJson(this);
 
   CARPDataFormat getCARPDataFormat() => CARP_DATA_FORMAT;
@@ -62,8 +59,7 @@ class GyroscopeDatum extends CARPDatum {
 /// A [Datum] that holds light intensity in Lux from the light sensor on the phone.
 @JsonSerializable(fieldRename: FieldRename.snake, includeIfNull: false)
 class LightDatum extends CARPDatum {
-  static CARPDataFormat CARP_DATA_FORMAT =
-      new CARPDataFormat(NameSpace.CARP_NAMESPACE, ProbeRegistry.LIGHT_MEASURE);
+  static CARPDataFormat CARP_DATA_FORMAT = new CARPDataFormat(NameSpace.CARP_NAMESPACE, ProbeRegistry.LIGHT_MEASURE);
 
   /// Intensity in Lux
   num avgLux;
@@ -71,15 +67,12 @@ class LightDatum extends CARPDatum {
   num minLux;
   num maxLux;
 
-  LightDatum({this.avgLux, this.stdLux, this.minLux, this.maxLux})
-      : super(includeDeviceInfo: false);
+  LightDatum({this.avgLux, this.stdLux, this.minLux, this.maxLux}) : super(includeDeviceInfo: false);
 
-  factory LightDatum.fromJson(Map<String, dynamic> json) =>
-      _$LightDatumFromJson(json);
+  factory LightDatum.fromJson(Map<String, dynamic> json) => _$LightDatumFromJson(json);
   Map<String, dynamic> toJson() => _$LightDatumToJson(this);
 
   CARPDataFormat getCARPDataFormat() => CARP_DATA_FORMAT;
 
-  String toString() =>
-      'light: {avgLux: $avgLux, stdLux: $stdLux, minLux: $minLux, avgLux: $maxLux}';
+  String toString() => 'light: {avgLux: $avgLux, stdLux: $stdLux, minLux: $minLux, maxLux: $maxLux}';
 }